+++ node-cli/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "axios": "^1.6.2",
     "commander": "^11.1.0",
+    "date-fns": "^2.30.0",
     "xml2js": "^0.6.2"
   }
 }
